Staging (#2)
* Upgrade

* Update deps

* rename file

* Rename file pt 2 (windows bad)

* another file rename

* Rename again

* Update Scribunto

* Add aptfile for lua bins

* update lua engine

* Update that one more time

* uppercase error again (scributno)

* uppercase

* Update cargo

* dont specify lua path

* OK re add this one

* Update page forms

* Fix AutoSitemap error

* Update popups

* Upgrade more deps

* Update AWS for images

* Increase max upload time

* Increase timeout even more

* Reduce

* Remove path

* Remove path

* update

* Add a .profile

+++ extensions/Cargo/includes/CargoBackLinks.php
@@ -8,8 +8,18 @@ if ( !defined( 'DB_PRIMARY' ) ) {
 }
 
 class CargoBackLinks {
+
+	/**
+	 * ParserOutput extension data key for backlinks.
+	 */
+	public const BACKLINKS_DATA_KEY = 'ext-cargo-backlinks';
+
 	public static function managePageDeletion( $pageId ) {
-		$page = \WikiPage::newFromID( $pageId );
+		if ( method_exists( MediaWikiServices::class, 'getWikiPageFactory' ) ) {
+			$page = MediaWikiServices::getInstance()->getWikiPageFactory()->newFromID( $pageId );
+		} else {
+			$page = \WikiPage::newFromID( $pageId );
+		}
 		$pageTitle = $page ? $page->getTitle() : null;
 		if ( $pageTitle ) {
 			$pageId = $pageTitle->getArticleID();
@@ -23,9 +33,14 @@ class CargoBackLinks {
 	}
 
 	public static function removeBackLinks( $pageId ) {
+		global $wgCargoIgnoreBacklinks;
+		if ( $wgCargoIgnoreBacklinks ) {
+			return;
+		}
+
 		$lb = MediaWikiServices::getInstance()->getDBLoadBalancer();
 		$dbw = $lb->getConnectionRef( DB_PRIMARY );
-		if ( $dbw->tableExists( 'cargo_backlinks' ) ) {
+		if ( $dbw->tableExists( 'cargo_backlinks' ) && !$dbw->isReadOnly() ) {
 			$dbw->delete( 'cargo_backlinks', [
 				'cbl_query_page_id' => $pageId
 			], __METHOD__ );
@@ -33,15 +48,23 @@ class CargoBackLinks {
 	}
 
 	public static function setBackLinks( $title, $resultsPageIds ) {
+		global $wgCargoIgnoreBacklinks;
+		if ( $wgCargoIgnoreBacklinks ) {
+			return;
+		}
+
 		$lb = MediaWikiServices::getInstance()->getDBLoadBalancer();
 		$dbw = $lb->getConnectionRef( DB_PRIMARY );
-		if ( !$dbw->tableExists( 'cargo_backlinks' ) ) {
+		if ( !$dbw->tableExists( 'cargo_backlinks' ) || $dbw->isReadOnly() ) {
 			return;
 		}
 		// Sanity check
 		$resultsPageIds = array_unique( $resultsPageIds );
+
 		$pageId = $title->getArticleID();
-		self::removeBackLinks( $pageId );
+		$dbw->delete( 'cargo_backlinks', [
+			'cbl_query_page_id' => $pageId
+		], __METHOD__ );
 
 		foreach ( $resultsPageIds as $resultPageId ) {
 			if ( $resultPageId ) {
@@ -54,6 +77,11 @@ class CargoBackLinks {
 	}
 
 	public static function purgePagesThatQueryThisPage( $resultPageId ) {
+		global $wgCargoIgnoreBacklinks;
+		if ( $wgCargoIgnoreBacklinks ) {
+			return;
+		}
+
 		$lb = MediaWikiServices::getInstance()->getDBLoadBalancer();
 		$dbr = $lb->getConnectionRef( DB_REPLICA );
 		if ( !$dbr->tableExists( 'cargo_backlinks' ) ) {
@@ -63,10 +91,21 @@ class CargoBackLinks {
 		$res = $dbr->select( 'cargo_backlinks',
 			[ 'cbl_query_page_id' ],
 			[ 'cbl_result_page_id' => $resultPageId ] );
+		if ( method_exists( MediaWikiServices::class, 'getWikiPageFactory' ) ) {
+			// MW 1.36+
+			$wikiPageFactory = MediaWikiServices::getInstance()->getWikiPageFactory();
+		} else {
+			$wikiPageFactory = null;
+		}
 		foreach ( $res as $row ) {
 			$queryPageId = $row->cbl_query_page_id;
 			if ( $queryPageId ) {
-				$page = \WikiPage::newFromID( $queryPageId );
+				if ( $wikiPageFactory !== null ) {
+					// MW 1.36+
+					$page = $wikiPageFactory->newFromID( $queryPageId );
+				} else {
+					$page = \WikiPage::newFromID( $queryPageId );
+				}
 				if ( $page ) {
 					$page->doPurge();
 				}
